shell: move muting logic to trx
Resolves #3333.
