+++ docs/tr2/CHANGELOG.md
@@ -7,6 +7,7 @@
 - added the ability to cycle UI tabs with sidestep keys (#3272)
 - added the ability to change the health bar color for allies, defaulting to green (#3005)
 - added the ability to skip consecutive credit images by holding the action / escape keys
+- added muting of game audio when the game window is not in focus
 - added dedicated British English translation (#3212)
 - added a new easter egg command
 - added a `/lighting` console command to let the player turn lighting system on/off
