+++ docs/tr2/README.md
@@ -376,6 +376,7 @@ However, you can easily download them manually from these urls:
 - added the ability to trigger different ambient tracks in custom levels, which will loop and be remembered between saves
 - added an option to continue playing music while in the inventory
 - added an option to control playing of certain animation sound effects such as doors when underwater
+- fixed game audio not muting when game is minimized
 - fixed music not playing with certain game versions
 - fixed the audio not being in sync when Lara strikes the gong in Ice Palace
 - fixed sound settings resuming the music
