+++ src/libtrx/include/libtrx/game/fmv.h
@@ -2,3 +2,6 @@
 
 extern bool FMV_Play(const char *path);
 extern bool FMV_IsPlaying(void);
+
+extern void FMV_Mute(void);
+extern void FMV_Unmute(void);
