+++ src/tr1/game/fmv.h
@@ -1,6 +1,3 @@
 #pragma once
 
 #include <libtrx/game/fmv.h>
-
-void FMV_Mute(void);
-void FMV_Unmute(void);
